src/test/java/com/leetcode/hashtable/easy/PalindromePermutation_266Test.java
package com.leetcode.hashtable.easy;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class PalindromePermutation_266Test {

  private final PalindromePermutation_266 solution = new PalindromePermutation_266();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(
          arguments("code", false), arguments("aab", true), arguments("carerac", true));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void canPermutePalindrome(String str, boolean expectedResult) {
    boolean actualResult = solution.canPermutePalindrome(str);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
